+++ service/collector/set.go
@@ -1,7 +1,7 @@
 package collector
 
 import (
-	infrastructurev1alpha2 "github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha2"
+	infrastructurev1alpha3 "github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha3"
 	"github.com/giantswarm/certs/v3/pkg/certs"
 	"github.com/giantswarm/exporterkit/collector"
 	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
@@ -14,7 +14,7 @@ type SetConfig struct {
 	K8sClient    k8sclient.Interface
 	Logger       micrologger.Logger
 
-	NewCommonClusterObjectFunc func() infrastructurev1alpha2.CommonClusterObject
+	NewCommonClusterObjectFunc func() infrastructurev1alpha3.CommonClusterObject
 }
 
 // Set is basically only a wrapper for the operator's collector implementations.