+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Introducing `v1alpha3` CR's.
+
 ## [3.9.0] - 2021-07-20
 
 ### Changed